move guardians methods to eos service refactor
this allows us to add higher level functions like add/remove the guard.seeds permission from the account. These are service level functions and should be in the service. Previously the getter was in http service.
