+++ lib/utils/extensions/response_extension.dart
@@ -1,0 +1,8 @@
+import 'dart:convert';
+import 'package:http/http.dart';
+
+extension ResponseExtension on Response {
+  dynamic parseJson() {
+    return json.decode(utf8.decode(this.bodyBytes));
+  }
+}
